create folder for view skill
